Update Localizable.strings
"App Addons" and "Carrier Blanks" weren't translated as we don't really have an accurate French translation (it would be "Ajout de fonctionnalités pour Applications" and "Modificateur de nom de fournisseur réseau", it might be too long to be usable and I can't translate it in a shorter way as it won't be enough understanable for most people). If you want me to do these changes, let me know.